[deploy] Bring back eternal diff checking for blackboxes (#294)
Long time ago, it was explicitly disabled for whiteboxes, but about a year ago
(in #23, git commit `7849fa9`) this logic was inverted, likely by mistake.

I do not see obvious issues with the feature, so I'm deleting this check at all.

Also, add small explanation before the diff ("The diff is still present:").